+++ src/main/java/roomescape/auth/JWTUtil.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
